Migrate remaining basic types to dc_bundle
Fixes #1453
