+++ crates/dc_bundle/src/legacy_definition/view/node_style.rs
@@ -14,5 +14,16 @@
  * limitations under the License.
  */
 
-pub mod node_style;
-pub mod view;
+use serde::{Deserialize, Serialize};
+
+#[derive(Copy, Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub enum Number {
+    Defined(f32),
+    Undefined,
+}
+
+impl Default for Number {
+    fn default() -> Self {
+        Self::Undefined
+    }
+}
