+++ crates/figma_import/src/toolkit_style.rs
@@ -26,7 +26,9 @@ use dc_bundle::legacy_definition::element::geometry::Size;
 use dc_bundle::legacy_definition::element::path::{LineHeight, Stroke};
 use dc_bundle::legacy_definition::element::variable::{ColorOrVar, NumOrVar};
 use dc_bundle::legacy_definition::interaction::pointer::PointerEvents;
-use dc_bundle::legacy_definition::layout::grid::{GridLayoutType, GridSpan};
+use dc_bundle::legacy_definition::layout::grid::{
+    Display, GridLayoutType, GridSpan, LayoutSizing, Overflow,
+};
 use dc_bundle::legacy_definition::layout::layout_style::LayoutStyle;
 use dc_bundle::legacy_definition::modifier::blend::BlendMode;
 use dc_bundle::legacy_definition::modifier::filter::FilterOp;
@@ -39,7 +41,7 @@ use dc_bundle::legacy_definition::plugin::meter_data::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::toolkit_layout_style::{Display, LayoutSizing, Number, Overflow};
+use dc_bundle::legacy_definition::view::node_style::Number;
 
 // These are the style properties that apply to text, so we can use them on subsections of
 // a longer string. We then assume that every style transition is a potential line break (and
