+++ crates/figma_import/src/tools/fetch_layout.rs
@@ -12,14 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    toolkit_layout_style::LayoutSizing, toolkit_schema::View, DesignComposeDefinition, Document,
-    ProxyConfig, ViewData,
-};
+use crate::{toolkit_schema::View, DesignComposeDefinition, Document, ProxyConfig, ViewData};
 /// Utility program to fetch a doc and serialize it to file
 use clap::Parser;
 use dc_bundle::legacy_definition::element::geometry::Dimension;
 use dc_bundle::legacy_definition::element::node::NodeQuery;
+use dc_bundle::legacy_definition::layout::grid::LayoutSizing;
 use layout::LayoutManager;
 use std::collections::HashMap;
 use std::io;
