+++ proto/definition/view/node_style.proto
@@ -21,7 +21,6 @@ import "definition/element/font.proto";
 import "definition/element/geometry.proto";
 import "definition/element/path.proto";
 import "definition/interaction/pointer.proto";
-import "definition/layout/display.proto";
 import "definition/layout/grid.proto";
 import "definition/modifier/blend.proto";
 import "definition/modifier/filter.proto";
